Change some ints to uints

+++ src/gpu/GridGenerator/grid/BoundaryConditions/Side.cpp
@@ -149,7 +149,7 @@ void Side::setQs(SPtr<Grid> grid, SPtr<BoundaryCondition> boundaryCondition, uin
         grid->transIndexToCoords(index, x, y, z);
 
         std::array<real, 3> coords = { x, y, z };
-        std::array<real, 3> neighborCoords = getNeighborCoordinates(grid.get(), coords, dir);
+        std::array<real, 3> neighborCoords = getNeighborCoordinates(grid.get(), coords, (uint)dir);
 
         correctNeighborForPeriodicBoundaries(grid.get(), coords, neighborCoords);
 
@@ -170,7 +170,7 @@ void Side::setQs(SPtr<Grid> grid, SPtr<BoundaryCondition> boundaryCondition, uin
     boundaryCondition->qs.push_back(qNode);
 }
 
-std::array<real, 3> Side::getNeighborCoordinates(Grid *grid, const std::array<real, 3> &coordinates, int direction) const
+std::array<real, 3> Side::getNeighborCoordinates(Grid *grid, const std::array<real, 3> &coordinates, uint direction) const
 {
     return { coordinates[0] + grid->getDirection()[direction * DIMENSION + 0] * grid->getDelta(),
              coordinates[1] + grid->getDirection()[direction * DIMENSION + 1] * grid->getDelta(),
